separando el enviromentMaager de los settings
