+++ composer.json
@@ -1,5 +1,6 @@
 {
 	"require":{
+		"php": ">=7.1.0",
 		"slim/slim":"3.*"
 	},
 	"autoload": {
