+++ src/settings.default.php
@@ -1,0 +1,16 @@
+<?php
+$env='DEV';
+
+$enviroments = [
+    'DEV'   =>     ['host'=>'127.0.0.1',
+                    'user' => 'batman',
+                    'pass' => 'batman',
+                    'dbname' => 'open_fixture'],
+    
+    'PROD'  =>      ['host'=>'127.0.0.1',
+                    'user' => 'batman',
+                    'pass' => 'batman',
+                    'dbname' => 'open_fixture']
+    
+];
+
